backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.3.4
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.1
django-admin-adminlte==1.0.5
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.2.1
gunicorn==21.2.0
h11==0.14.0
idna==3.7
inflection==0.5.1
kombu==5.3.7
mysqlclient==2.2.0
outcome==1.3.0.post0
packaging==24.0
prompt-toolkit==3.0.43
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
selenium==4.9.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wsproto==1.2.0
